[iOS][STG] Don't show delete group in incognito
This CL hides the delete group context menu action in incogntio.
The group view entry should also be updated in a follow-up.

Bug: 352297050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5701960
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326767}

+++ ios/chrome/browser/saved_tab_groups/model/ios_tab_group_sync_util.mm
@@ -82,7 +82,8 @@ LocalTabInfo GetLocalTabInfo(WebStateList* web_state_list,
 void CloseTabGroupLocally(const TabGroup* tab_group,
                           WebStateList* web_state_list,
                           TabGroupSyncService* sync_service) {
-  if (sync_service->GetGroup(tab_group->tab_group_id())) {
+  // `sync_service` is nullptr in incognito.
+  if (sync_service && sync_service->GetGroup(tab_group->tab_group_id())) {
     sync_service->RemoveLocalTabGroupMapping(tab_group->tab_group_id());
   }
   CloseAllWebStatesInGroup(*web_state_list, tab_group,
